+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2025-03-21
+
+### Added
+
+- `Config` 新增 `internal.devTools` 配置项
+
+### Fixed
+
+- 修复部分环境下 `DevTools` 初始化失败引发的崩溃 #13
+
 ## [0.5.0] - 2025-02-19
 
 ### Added